refactor: implement domain-driven architecture and eliminate dataset package
## Major Changes

### Architecture Refactoring
- Implemented domain-driven design (DDD) with package-oriented structure
- Each business domain (accounts, packs, inventories, images) now has its own package
- Eliminated centralized `dataset` package to avoid tight coupling

### Package Structure
- **pkg/accounts/**: User accounts and authentication types
- **pkg/packs/**: Pack management types
- **pkg/inventories/**: Inventory item types
- **pkg/images/**: Image management types and interfaces
- **pkg/apitypes/**: Shared HTTP response types (OkResponse, ErrorResponse)
- **pkg/config/**: Configuration types including MailServer

### Domain-Specific Helpers
Moved helper functions to their respective domain packages:
- `FindUserIDByUsername()` → pkg/accounts/
- `FindPackIDByPackName()` → pkg/packs/
- `FindItemIDByItemName()` → pkg/inventories/

### Type Organization
Each package now contains:
- `types.go`: Domain-specific type definitions
- `handlers.go` or `<package>.go`: HTTP handlers and business logic
- `service.go`: Public service functions (where applicable)
- `repository.go`: Private data access functions (where applicable)
- `testdata.go`: Test fixtures using domain types

### Circular Dependency Resolution
- Moved `MailServer` type to `pkg/config/` to avoid circular dependency between helper and config
- Test data files now use domain-specific types directly
- Helper package only contains generic utilities

### Documentation
- Updated agents/PATTERNS.md with comprehensive DDD architecture guidelines
- Added examples for proper package organization
- Documented helper function placement patterns
- Added Swagger documentation patterns with domain-specific types

### Testing
- All test files updated to use domain-specific types
- Test data setup functions use domain types
- Maintained test coverage while improving code organization

## Benefits
- ✅ Better separation of concerns
- ✅ Clearer domain boundaries
- ✅ Eliminated tight coupling via centralized dataset package
- ✅ No circular dependencies
- ✅ Easier to maintain and extend
- ✅ Follows Go best practices for package organization

## Validation
- ✅ All packages build successfully
- ✅ Linter passes with 0 issues
- ✅ Domain-specific types properly encapsulated
