+++ pkg/helper/helper_test.go
@@ -6,7 +6,6 @@ import (
 	"testing"
 
 	"github.com/Angak0k/pimpmypack/pkg/config"
-	"github.com/Angak0k/pimpmypack/pkg/dataset"
 )
 
 func TestMain(m *testing.M) {
@@ -54,116 +53,6 @@ func TestConvertWeightUnit(t *testing.T) {
 			output := ConvertWeightUnit(tc.input)
 			if output != tc.expected {
 				t.Errorf("ConvertWeightUnit(%s): expected %s, got %s", tc.input, tc.expected, output)
-			}
-		})
-	}
-}
-
-func TestFindUserIDByUsername(t *testing.T) {
-	// Define test cases
-	testCases := []struct {
-		users []dataset.User
-		// Input users
-		username string // Input username
-		expected uint   // Expected output
-	}{
-		{
-			[]dataset.User{{ID: 1, Username: "user1"}, {ID: 2, Username: "user2"}},
-			"user1",
-			1,
-		},
-		{
-			[]dataset.User{{ID: 1, Username: "user1"}, {ID: 2, Username: "user2"}},
-			"user2",
-			2,
-		},
-		{
-			[]dataset.User{{ID: 1, Username: "user1"}, {ID: 2, Username: "user2"}},
-			"user3",
-			0,
-		},
-	}
-
-	// Iterate through test cases
-	for _, tc := range testCases {
-		t.Run(tc.username, func(t *testing.T) {
-			output := FindUserIDByUsername(tc.users, tc.username)
-			if output != tc.expected {
-				t.Errorf("FinUserIDByUsername(%v, %s): expected %d, got %d", tc.users, tc.username, tc.expected, output)
-			}
-		})
-	}
-}
-
-func TestFindPackIDByPackName(t *testing.T) {
-	// Define test cases
-	testCases := []struct {
-		packs    dataset.Packs
-		packname string // Input packname
-		expected uint   // Expected output
-	}{
-		{
-			dataset.Packs{{ID: 1, PackName: "pack1"}, {ID: 2, PackName: "pack2"}},
-			"pack1",
-			1,
-		},
-		{
-			dataset.Packs{{ID: 1, PackName: "pack1"}, {ID: 2, PackName: "pack2"}},
-			"pack2",
-			2,
-		},
-		{
-			dataset.Packs{{ID: 1, PackName: "pack1"}, {ID: 2, PackName: "pack2"}},
-			"pack3",
-			0,
-		},
-	}
-
-	// Iterate through test cases
-	for _, tc := range testCases {
-		t.Run(tc.packname, func(t *testing.T) {
-			output := FindPackIDByPackName(tc.packs, tc.packname)
-			if output != tc.expected {
-				t.Errorf("FinPackIDByPackName(%v, %s): expected %d, got %d", tc.packs, tc.packname, tc.expected, output)
-			}
-		})
-	}
-}
-
-func TestFindItemIDByItemName(t *testing.T) {
-	// Define test cases
-	testCases := []struct {
-		inventories dataset.Inventories
-		itemname    string // Input itemname
-		expected    uint   // Expected output
-	}{
-		{
-			dataset.Inventories{{ID: 1, ItemName: "item1"}, {ID: 2, ItemName: "item2"}},
-			"item1",
-			1,
-		},
-		{
-			dataset.Inventories{{ID: 1, ItemName: "item1"}, {ID: 2, ItemName: "item2"}},
-			"item2",
-			2,
-		},
-		{
-			dataset.Inventories{{ID: 1, ItemName: "item1"}, {ID: 2, ItemName: "item2"}},
-			"item3",
-			0,
-		},
-	}
-
-	// Iterate through test cases
-	for _, tc := range testCases {
-		t.Run(tc.itemname, func(t *testing.T) {
-			output := FindItemIDByItemName(tc.inventories, tc.itemname)
-			if output != tc.expected {
-				t.Errorf("FinItemIDByItemName(%v, %s): expected %d, got %d",
-					tc.inventories,
-					tc.itemname,
-					tc.expected,
-					output)
 			}
 		})
 	}
